requirements.txt
# Базовые зависимости
python-telegram-bot>=20.7
matplotlib
pandas
pytz
python-dotenv